씬나누기 카메라 완료

+++ PortfolioGame/PortfolioGame/MainGame/main_game.h
@@ -24,7 +24,7 @@ private:
 
 
 	int _iFPS;
-	DWORD _dwFPSTime;
+	ULONGLONG _dwFPSTime;
 	TCHAR _szFPS[32];
 };
