+++ PortfolioGame/PortfolioGame/PublicHeader/all_extern.h
@@ -1,6 +1,6 @@
 #pragma once
 
 extern HWND _hWnd;
-const DWORD WindowCX = 1024;
-const DWORD WindowCY = 768;
+const LONG WindowCX = 1024;
+const LONG WindowCY = 768;
 const uint8_t MaxLayer = 7;
